Add storage file for config.
This commit adds the capability for the script to store the network segments in
a csv (both name of segment and names of each target).
If the file exists, the contents are read. If it doesn't exist, it is created
and the input from the user is written to it.

Signed-off-by: = <supertaxidriver@gmail.com>